Add support for custom Bedrock transports
